Доработка кода (pep-8)
