+++ lab3/mai.py
@@ -3,8 +3,7 @@ import argparse
 import cv2
 import matplotlib.pyplot as plt
 
-import numpy as np
-import numpy.typing as npt
+from Work_with_images import write_image,gistogramm,task_binary_image
 
 
 def parsing() -> argparse.Namespace:
@@ -17,51 +16,6 @@ def parsing() -> argparse.Namespace:
     parser.add_argument('path_to_binary_image', type = str, help='path to the binary_output image')
     args = parser.parse_args()
     return args
-
-
-def write_image(input_image, greyscale) -> np.ndarray:
-    """
-    Запись нового изображения в файл
-    :param input_image: входное изображение
-    :param greyscale: аргумент для обработки серых изображений
-    :return:  Mat | ndarray[Any, dtype] | ndarray
-    """
-    image = cv2.imread(input_image, greyscale)
-    if image is None:
-        raise FileExistsError
-    return image
-
-
-def gistogramm(image) -> None:
-    """
-    Функция для построения и отображения гистограммы изображения
-    :param image: Изображение для построение гистограммы
-    :return: None
-    """
-    plt.figure()
-    plt.title("Grayscale Histogram")
-    plt.xlabel("Bins")
-    plt.ylabel("# of Pixels")
-    # Аргументы: 1-само изображение 2-канал(серый,зеленый,синий) 3-mask(срез),
-    # 4-количество интервалов, 5-диапозон на графике
-    hist = cv2.calcHist([image], [0], None, [256], [0, 256])
-    plt.plot(hist)
-    plt.xlim([0, 256])
-    plt.show()
-
-
-def task_binary_image(image) -> npt.array():
-    """
-    Преобразование в бинарное изображение
-    :param image:
-    :return: многомерный массив
-    """
-    '''
-    Аргументы: 1-само изображение 2-пороговое значение(ниже пиксели черные, выше-белые)
-    3-значение присвоенное пикселям прошедшим порог 4-бинарное преобразование
-    '''
-    _, binary_image = cv2.threshold(image, 127, 255, cv2.THRESH_BINARY)
-    return binary_image
 
 
 def show_all(orig_image, grey_image, binary_image) -> None:
